+++ SpellWork/DBC/Structures/SpellCategoryEntry.cs
@@ -5,12 +5,12 @@ namespace SpellWork.DBC.Structures
     public class SpellCategoryEntry
     {
         [Index(true)]
-        public uint ID;
+        public int ID;
         public string Name;
-        public sbyte Flags;
+        public int Flags;
         public byte UsesPerWeek;
         public sbyte MaxCharges;
         public int ChargeRecoveryTime;
-        public int TypeMask;
+        public uint TypeMask;
     };
 }
